+++ src/commonTest/kotlin/kotools/types/experimental/BoundTest.kt
@@ -18,8 +18,8 @@ class InclusiveBoundTest {
     @Test
     fun equals_should_pass_with_an_inclusive_bound_having_the_same_value() {
         // GIVEN
-        val x: InclusiveBound<Int> = InclusiveBound(Random.nextInt())
-        val y: InclusiveBound<Int> = InclusiveBound(x.value)
+        val x: InclusiveBound<Int> = InclusiveBound of Random.nextInt()
+        val y: InclusiveBound<Int> = InclusiveBound of x.value
         // WHEN & THEN
         "The equals operation should be reflexive.".let {
             assertEquals(expected = x, actual = x, message = it)
@@ -30,7 +30,7 @@ class InclusiveBoundTest {
             assertEquals(expected = y, actual = x, message = it)
         }
         "The equals operation should be transitive.".let {
-            val z: InclusiveBound<Int> = InclusiveBound(y.value)
+            val z: InclusiveBound<Int> = InclusiveBound of y.value
             assertEquals(expected = x, actual = y, message = it)
             assertEquals(expected = y, actual = z, message = it)
             assertEquals(expected = x, actual = z, message = it)
@@ -46,8 +46,8 @@ class InclusiveBoundTest {
     @Test
     fun equals_should_fail_with_an_inclusive_bound_having_another_value() {
         // GIVEN
-        val x: InclusiveBound<Int> = InclusiveBound(Random.nextInt())
-        val y: InclusiveBound<Int> = InclusiveBound(Random.nextInt())
+        val x: InclusiveBound<Int> = InclusiveBound of Random.nextInt()
+        val y: InclusiveBound<Int> = InclusiveBound of Random.nextInt()
         // WHEN & THEN
         assertNotEquals(illegal = x, actual = y)
     }
@@ -56,7 +56,7 @@ class InclusiveBoundTest {
     fun toString_should_return_the_string_representation_of_its_value() {
         // GIVEN
         val value: Int = Random.nextInt()
-        val bound: InclusiveBound<Int> = InclusiveBound(value)
+        val bound: InclusiveBound<Int> = InclusiveBound of value
         // WHEN
         val result = "$bound"
         // THEN